db/etl/requirements.txt
psycopg2-binary>=2.9
pandas>=1.5
SQLAlchemy>=1.4
openpyxl>=3.0.0